D3: decompose counting_cog into cogs/counting/ + views/counting/
Reduces cogs/counting_cog.py from 1380 → ~680 LOC (~51% reduction)
by extracting the parsing pipeline, game-mode arithmetic, and the
admin hub view into dedicated modules. The cog now hosts only
Discord plumbing: commands, the on_message listener, lifecycle, and
permission helpers.

New layout:
  disbot/cogs/counting/
  ├── __init__.py
  ├── _constants.py    NUMBER_WORDS_SET / PHRASE_NUMBER_MAPPING /
  │                    ORDINAL_MAPPING / ROMAN_NUMERAL_MAPPING /
  │                    EMOJI_NUMBER_MAPPING / OPERATORS /
  │                    OPERATOR_MAPPING + compiled regex + word_to_num
  ├── parsing.py       parse_message, parse_number_word,
  │                    split_concatenated_numbers, roman_to_int,
  │                    eval_expr, safe_eval, _eval_ast — all pure
  │                    functions, no instance state
  └── game_logic.py    calculate_expected_count, is_prime — pure

  disbot/views/counting/
  ├── __init__.py
  └── hub_panel.py     _CountingHubView

Behaviour preservation:
- Every parser branch and game-mode arithmetic line is preserved
  verbatim; only ``self`` was removed and dicts/regex were lifted to
  module-level.
- on_message listener delegates to parsing.parse_message and
  game_logic.calculate_expected_count / is_prime; the lock contract,
  the tasks.spawn save pattern, and every embed message string are
  unchanged.

Back-compat re-exports:
- cogs.counting_cog._word_to_num (alias for word_to_num) so any test
  or external import keeps resolving.
- cogs.counting_cog._CountingHubView (re-exported from views.counting).

Persistence rule honoured: no SUBSYSTEM string was renamed, no panel
anchor or session row references changed, and the JSONB blob shape
in counting_state is identical.

Verification: ruff/black/isort/mypy clean; pytest 436/436 green.
